# docs/swagger.yaml
basePath: /
definitions:
  httpserver.ServiceHours:
    properties:
      end:
        example: "23:00"
        type: string
      start:
        example: "18:00"
        type: string
      tz:
        example: America/Sao_Paulo
        type: string
    type: object
  httpx.ErrorResponse:
    properties:
      error:
        example: Grupo de 6 pessoas excede a capacidade da mesa (4).
        type: string
    type: object
  reservation.Reservation:
    properties:
      created_at:
        type: string
      customer_name:
        example: Maria Silva
        type: string
      customer_phone:
        example: "11999998888"
        type: string
      ends_at:
        type: string
      id:
        type: string
      party_size:
        example: 4
        type: integer
      starts_at:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/reservation.Status'
        example: confirmed
      table_ids:
        items:
          type: string
        type: array
    type: object
  reservation.Status:
    enum:
    - confirmed
    - cancelled
    type: string
    x-enum-varnames:
    - StatusConfirmed
    - StatusCancelled
  reservation.TableAvailability:
    properties:
      capacity:
        example: 4
        type: integer
      free_windows:
        items:
          $ref: '#/definitions/reservation.Window'
        type: array
      table_id:
        type: string
      table_name:
        example: Mesa 04
        type: string
    type: object
  reservation.Window:
    properties:
      ends_at:
        type: string
      starts_at:
        type: string
    type: object
  reservation.createRequest:
    properties:
      customer_name:
        example: Maria Silva
        type: string
      customer_phone:
        example: "11999998888"
        type: string
      ends_at:
        example: "2026-08-01T21:00:00-03:00"
        type: string
      party_size:
        example: 4
        type: integer
      starts_at:
        example: "2026-08-01T19:00:00-03:00"
        type: string
      table_ids:
        description: |-
          Ausente, null ou vazio → "escolha a mesa por mim" (heurística automática).
          Uma mesa  → override manual.
          Várias    → COMBINAÇÃO: o staff empurrou as mesas, o sistema registra e
                      protege cada uma delas com a EXCLUDE (Fase 3a).

          Era `table_id` (uuid único) até a Fase 3a. Mudança quebrada de contrato,
          feita agora porque o frontend ainda não existe — depois dele, custaria o
          dobro.
        items:
          type: string
        type: array
    type: object
  table.Table:
    properties:
      capacity:
        type: integer
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
  table.createRequest:
    properties:
      capacity:
        example: 4
        type: integer
      name:
        example: Mesa 12
        type: string
    type: object
  table.updateRequest:
    properties:
      capacity:
        example: 6
        type: integer
      is_active:
        example: false
        type: boolean
      name:
        example: Mesa 12
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend de reservas de mesas de um restaurante único. Staff cria reservas;
    o sistema aloca mesa automaticamente respeitando capacidade e evitando sobreposição
    de horário.
  title: API de Reservas de Restaurante
  version: "1.0"
paths:
  /availability:
    get:
      description: Devolve, para cada mesa ativa, as janelas livres no dia dentro
        do horário de funcionamento. É a resposta para "quais mesas posso combinar
        às 20h?" — o cliente filtra a grade por contenção de intervalo, sem refazer
        o cálculo. Mesas inativas não aparecem.
      parameters:
      - description: Dia, AAAA-MM-DD, no fuso do restaurante
        example: "2026-07-20"
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/reservation.TableAvailability'
            type: array
        "400":
          description: date ausente ou malformado
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Grade de disponibilidade do dia (todas as mesas)
      tags:
      - mesas
  /reservations:
    get:
      description: Filtros opcionais e combináveis. `date` é interpretado no fuso
        do restaurante, não em UTC.
      parameters:
      - description: Dia da reserva, AAAA-MM-DD, no fuso do restaurante
        example: "2026-08-01"
        in: query
        name: date
        type: string
      - description: Filtra por mesa (UUID)
        in: query
        name: table_id
        type: string
      - description: confirmed ou cancelled
        enum:
        - confirmed
        - cancelled
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/reservation.Reservation'
            type: array
        "400":
          description: Parâmetro de filtro inválido
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Lista reservas
      tags:
      - reservas
    post:
      consumes:
      - application/json
      description: '`table_ids` vazio ou omitido: o sistema aloca automaticamente
        a menor mesa livre que comporte o grupo. Com uma mesa: usa exatamente aquela.
        Com várias: registra uma COMBINAÇÃO (o staff empurrou as mesas), validando
        a capacidade somada e protegendo cada mesa individualmente contra sobreposição.'
      parameters:
      - description: Dados da reserva
        in: body
        name: reserva
        required: true
        schema:
          $ref: '#/definitions/reservation.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/reservation.Reservation'
        "400":
          description: Dados inválidos, fora do expediente, mesa inativa/inexistente/repetida,
            ou grupo maior que a capacidade
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Mesa(s) pedida(s) ocupada(s), ou nenhuma mesa disponível no
            horário
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Cria uma reserva
      tags:
      - reservas
  /reservations/{id}:
    delete:
      description: 'Soft delete: a linha permanece com status `cancelled` e o horário
        é liberado para reuso. Idempotente — cancelar duas vezes devolve 204.'
      parameters:
      - description: ID da reserva (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Cancelada
        "400":
          description: ID não é um UUID
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Reserva não encontrada
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Cancela uma reserva
      tags:
      - reservas
    get:
      parameters:
      - description: ID da reserva (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.Reservation'
        "400":
          description: ID não é um UUID
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Reserva não encontrada
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Detalha uma reserva
      tags:
      - reservas
    patch:
      consumes:
      - application/json
      description: 'Cancela a reserva atual e cria uma nova com os dados enviados,
        numa transação só. A reserva devolvida tem **id novo** — editar é remarcar,
        não mutar a linha. O corpo é o mesmo do POST: `table_ids` vazio realoca automaticamente,
        com uma ou mais mesas usa exatamente essas. Se a nova alocação colidir, nada
        muda: a reserva original permanece.'
      parameters:
      - description: ID da reserva a editar (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Novos dados da reserva
        in: body
        name: reserva
        required: true
        schema:
          $ref: '#/definitions/reservation.createRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.Reservation'
        "400":
          description: Dados inválidos, fora do expediente, mesa inativa/inexistente/repetida,
            ou grupo maior que a capacidade
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Reserva não encontrada ou já cancelada
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Mesa(s) pedida(s) ocupada(s), ou nenhuma mesa disponível no
            horário
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Edita (remarca) uma reserva
      tags:
      - reservas
  /service-hours:
    get:
      description: 'Horário de funcionamento e fuso usados por toda a API: a validação
        de `starts_at` na criação de reserva e o cálculo das janelas livres. O frontend
        lê daqui em vez de guardar uma cópia.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpserver.ServiceHours'
      summary: Expediente do restaurante
      tags:
      - restaurante
  /tables:
    get:
      description: Lista as mesas. Sem o parâmetro `active`, devolve todas — ativas
        e inativas.
      parameters:
      - description: Filtra por mesas ativas (true) ou inativas (false). Omitido =
          sem filtro
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/table.Table'
            type: array
        "400":
          description: Parâmetro active não é booleano
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Lista mesas
      tags:
      - mesas
    post:
      consumes:
      - application/json
      description: Cadastra uma mesa nova. O nome é único no restaurante.
      parameters:
      - description: Dados da mesa
        in: body
        name: mesa
        required: true
        schema:
          $ref: '#/definitions/table.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/table.Table'
        "400":
          description: Corpo inválido, nome vazio ou capacidade fora do intervalo
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Já existe mesa com esse nome
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Cria uma mesa
      tags:
      - mesas
  /tables/{id}:
    get:
      parameters:
      - description: ID da mesa (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/table.Table'
        "400":
          description: ID não é um UUID
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Mesa não encontrada
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Detalha uma mesa
      tags:
      - mesas
    patch:
      consumes:
      - application/json
      description: 'Atualização parcial: só os campos enviados são alterados. Enviar
        `{"is_active": false}` desativa a mesa.'
      parameters:
      - description: ID da mesa (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Campos a alterar (ao menos um)
        in: body
        name: mesa
        required: true
        schema:
          $ref: '#/definitions/table.updateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/table.Table'
        "400":
          description: Corpo vazio, ID inválido ou valor fora do intervalo
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Mesa não encontrada
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Já existe mesa com o novo nome, ou a mesa tem reservas confirmadas
            e não pode ser desativada
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Atualiza uma mesa
      tags:
      - mesas
  /tables/{id}/availability:
    get:
      description: Devolve os intervalos livres da mesa dentro do horário de funcionamento
        do restaurante, no fuso dele. Mesa inativa devolve lista vazia.
      parameters:
      - description: ID da mesa (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Dia, AAAA-MM-DD, no fuso do restaurante
        example: "2026-07-20"
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/reservation.Window'
            type: array
        "400":
          description: ID inválido, date ausente/malformado, ou mesa inexistente
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Janelas livres de uma mesa no dia
      tags:
      - mesas
swagger: "2.0"